% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{authorMeshTree}
\alias{authorMeshTree}
\title{Extract the MeSH tree for a specific author}
\usage{
authorMeshTree(auID, dbInfo)
}
\arguments{
\item{auID}{Author ID (stored in the author table of the DB)}

\item{dbInfo}{Path to the ColabNet database. Can be left blank if setup with dbConn()}
}
\value{
A list with two data frames:
\itemize{
\item auTree: The full meshtree for the author (including intermediate nodes)
\item arMesh: The set of MeSH terms actually referenced in the author's articles
}
}
\description{
Extract the MeSH tree for a specific author
}
