% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePubmed.R
\name{authorPublicationInfo}
\alias{authorPublicationInfo}
\title{Get author names, papers, co-authors and affiliations for a specific researcher}
\usage{
authorPublicationInfo(firstName, lastName, n = -1)
}
\arguments{
\item{firstName}{First name of author of interest}

\item{lastName}{Last name of author of interest}

\item{n}{Number of papers to fetch from Pubmed}
}
\value{
List with 5 data frames
author: Names the author used in papers
articles List of author's articles on Pubmed
coAuthors Co-authors on each paper
affiliations Affiliation of each author
meshDescriptors MeSH term descriptor associated with each paper
meshQualifiers MeSH term qualifier associated with each paper
}
\description{
Get author names, papers, co-authors and affiliations for a specific researcher
}
