% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{authorSimScore}
\alias{authorSimScore}
\title{Score all author pairs based on shared MeSH terms}
\usage{
authorSimScore(difftree, minLevel = 1)
}
\arguments{
\item{difftree}{a diffTree dataframe returned by diffTree(pruneDuplicates = T)}

\item{minLevel}{(Default 1, root) The minimum level to consider a term shared}
}
\value{
A dataset that can be used to create a (plotly) Treemap
}
\description{
Score all author pairs based on shared MeSH terms
}
