% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{branchID}
\alias{branchID}
\title{Find and label linear branches (single chain of nodes)
These will be labelled with the same branchID, as they can be collapse}
\usage{
branchID(node, parent)
}
\arguments{
\item{node}{Vector of node IDs}

\item{parent}{Vector of parent ID for each node above}
}
\value{
A list with two elements
\itemize{
\item branchID: The new branch ID for each node
\item treemapVal: The value used to scale a treemap and force all cells to
become roughly the same size
}
}
\description{
Meanwhile, a treemapVal for each leaf node will be calculated such that
that when rendered as a treemap all cells should roughly be of the same size
}
