% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{busyModal}
\alias{busyModal}
\title{Show a screen overlay when busy to prevent other actions
By default an animation is played to indicate the server is busy.
removeModal has to be called separately}
\usage{
busyModal(session, message = NULL, title = NULL, showAnimation = T)
}
\arguments{
\item{session}{A Shiny session object}

\item{message}{(Optional) Text to add to the modal}

\item{title}{(Optional) Title of the modal}

\item{showAnimation}{Default = T. Show a busy animation}
}
\value{
A modal pop-up window
}
\description{
Show a screen overlay when busy to prevent other actions
By default an animation is played to indicate the server is busy.
removeModal has to be called separately
}
