% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{checkTreeNums}
\alias{checkTreeNums}
\title{Check if MeSH tree numbers are valid}
\usage{
checkTreeNums(treenums, output = "errorOnly")
}
\arguments{
\item{treenums}{vector of MeSH tree numbers}

\item{output}{"errorOnly" (Default) will only raise error, not output.
Alternatively "bool" will return TRUE / FALSE for no errors / errors respectively
"vector" will return a list of TRUE / FALSE with the check for each number}
}
\value{
Dependent on output argument
}
\description{
Check if MeSH tree numbers are valid
}
