% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{collapseTree}
\alias{collapseTree}
\title{Collapse a Mesh Tree based on the branchIDs
All nodes in the same linear branch are merged to reduce unnecessary nesting}
\usage{
collapseTree(meshtree, sep = " -> <br>")
}
\arguments{
\item{meshtree}{Data frame generated by dbMeshTree()}

\item{sep.}{Default = " -> \if{html}{\out{<br>}}". Separator to use for MeSH terms when
collapsing them across nodes}
}
\value{
Dataframe containing the collapsed MeSH tree
}
\description{
Collapse a Mesh Tree based on the branchIDs
All nodes in the same linear branch are merged to reduce unnecessary nesting
}
