% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{copubGraphElements}
\alias{copubGraphElements}
\title{Create graph elements (nodes, edges) from co-publication data frame}
\usage{
copubGraphElements(articleInfo)
}
\arguments{
\item{articleInfo}{A data frame representing co-publications with at least
the following columns
\itemize{
\item arID
\item auID
\item lastName
\item firstName
\item year: year of article publication
\item month: month of article publication.
}}
}
\value{
a list of nodes an edges used to generate graphs / stats
}
\description{
Create graph elements (nodes, edges) from co-publication data frame
}
