% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{copubGraphStats}
\alias{copubGraphStats}
\title{Get statistics about the co-publication network}
\usage{
copubGraphStats(graphElements)
}
\arguments{
\item{graphElements}{A list of graph elements generated by copubGraphElements()}
}
\value{
A list of statistics describing the graph
}
\description{
Get statistics about the co-publication network
}
