% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{copubTrendInfo}
\alias{copubTrendInfo}
\title{Calculate the stats trends for a co-publication network over time}
\usage{
copubTrendInfo(articleInfo, windowSize = 5)
}
\arguments{
\item{articleInfo}{Data frame with articleInfo}

\item{windowSize}{Default = 5. The size of the sliding window to calculate stats}
}
\value{
list with two elements
\itemize{
\item trends: A data frame with the trend stats over time
\item plot: a ggplot showing the stats over time in facets
}
}
\description{
Calculate the stats trends for a co-publication network over time
}
