% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddArticles}
\alias{dbAddArticles}
\title{Add authors to the database}
\usage{
dbAddArticles(authorPublicationInfo, colabDB = colabDB)
}
\arguments{
\item{authorPublicationInfo}{List of data frames geneated by ncbi_authorPublicationInfo()}
}
\value{
Data is inserted into the DB and list of new, modified an unmodified author IDs (auID) is returned
}
\description{
Add authors to the database
}
