% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddAuthorPublications}
\alias{dbAddAuthorPublications}
\title{Add authors to the database}
\usage{
dbAddAuthorPublications(
  authorPublications,
  matchOnFirst = F,
  flagUpdate = T,
  dbInfo,
  commit = T
)
}
\arguments{
\item{authorPublications}{List of data frames geneated by ncbi_publicationDetails()}

\item{flagUpdate}{Default= T, set an update flag in the DB so the app will refresh once completed}

\item{dbInfo}{Connection to a Colabnet database}

\item{commmit}{(Default = TRUE). Commit changes to the database automatically}
}
\value{
Data is inserted into the DB returning a list the author's aritcles (arID and PMID)
}
\description{
Add authors to the database
}
