% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbDeleteArticle}
\alias{dbDeleteArticle}
\title{Delete specified articles and any associated authors / author names
Useful when an article has been incorrectly associated with an author for
example because of name overlap}
\usage{
dbDeleteArticle(arIDs, conn, commit = T)
}
\arguments{
\item{arIDs}{Vector of arIDs to delete (as found in article table)}

\item{conn}{Connection to a Colabnet database}

\item{commit}{(Default = TRUE). Commit changes to the database automatically}
}
\value{
Dataframe of articles and authors/author names that have been deleted
}
\description{
Delete specified articles and any associated authors / author names
Useful when an article has been incorrectly associated with an author for
example because of name overlap
}
