% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbFinish}
\alias{dbFinish}
\title{Finish the current DB operation and disconnect if needed}
\usage{
dbFinish(conn)
}
\arguments{
\item{conn}{A database connection}
}
\value{
Nothing
}
\description{
Finish the current DB operation and disconnect if needed
}
