% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbFlagUpdate}
\alias{dbFlagUpdate}
\title{Tell the colabNet app an update has been made and it needs to refresh}
\usage{
dbFlagUpdate(action, conn, commit = T)
}
\arguments{
\item{action}{(Integer) The code of the action
\itemize{
\item 0 = new database init
\item 1 = add data
\item 2 = delete data
}}

\item{conn}{Connection to a Colabnet database}

\item{commit}{(Default = TRUE). Commit changes to the database automatically}
}
\value{
The uID from inserted row into updateData
}
\description{
Tell the colabNet app an update has been made and it needs to refresh
}
