% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Get a ColabNet database connection}
\usage{
dbGetConn(dbInfo)
}
\arguments{
\item{dbInfo}{Connection to a Colabnet database}
}
\value{
Connection to the ColabNet database
}
\description{
Get a ColabNet database connection
}
