% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{dbPaperMesh}
\alias{dbPaperMesh}
\title{Get all of the MeSH terms from papers by a set of authors}
\usage{
dbPaperMesh(auIDs, roots, conn)
}
\arguments{
\item{auIDs}{Vector auIDs}

\item{roots}{(Optional) Vector of single letter representing the tree roots
to include. If not specified, all categories are returned}

\item{conn}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe with paper Mesh terms for authors
}
\description{
Get all of the MeSH terms from papers by a set of authors
}
