% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbSetup}
\alias{dbSetup}
\title{Setup the ColabNet database for the current session}
\usage{
dbSetup(dbInfo, newDBMsg = T, checkSchema = F, returnConn = F)
}
\arguments{
\item{dbInfo}{Path to ColabNet Database. If DB does not exist it will be created}

\item{newDBMsg}{(Default, TRUE) Output a message when a new database was created}

\item{checkSchema}{(Default, FALSE) Check the schema of an existing datbase against the reference}

\item{returnConn}{(Default, FALSE) By default this function will save the connection to the database
internally to be used by other functions. If set to TRUE, a connection object is returned
and needs to be closed manually using dbDisconnect()}
}
\value{
Nothing (default) or a connection to the ColabNet datbase if returnConn = T
}
\description{
Setup the ColabNet database for the current session
}
