% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{diffTree}
\alias{diffTree}
\title{Compare two author MeSH trees and find overlapping areas}
\usage{
diffTree(auIDs, pruneDuplicates = F, dbInfo)
}
\arguments{
\item{auIDs}{Vector author IDs (stored in the author table of the DB)}

\item{pruneDuplicates}{Prune parts of the MeSH tree that contain duplicated terms}

\item{dbInfo}{Path to the ColabNet database. Can be left blank if setup with dbConn()}
}
\value{
A data frame with the MeSH tree information of both authors and their overlap
}
\description{
Compare two author MeSH trees and find overlapping areas
}
