% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{entrezFetch}
\alias{entrezFetch}
\title{Temp wrapper around entrez_fetch because of errors in fetching data}
\usage{
entrezFetch(..., tries = 5)
}
\arguments{
\item{...}{All function params}

\item{tries}{(Default = 3) Times to try the API  call}
}
\value{
entrez_fetch output
}
\description{
Temp wrapper around entrez_fetch because of errors in fetching data
}
