% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{entrezSearch}
\alias{entrezSearch}
\title{Temp wrapper around entrez_search because of errors in fetching data}
\usage{
entrezSearch(..., tries = 5)
}
\arguments{
\item{...}{All function params}

\item{tries}{(Default = 3) Times to try the API  call}
}
\value{
entrez_search output
}
\description{
Temp wrapper around entrez_search because of errors in fetching data
}
