% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{entrezSummary}
\alias{entrezSummary}
\title{Temp wrapper around entrez_summary because of errors in fetching data}
\usage{
entrezSummary(..., tries = 10)
}
\arguments{
\item{...}{All function params}

\item{tries}{(Default = 3) Times to try the API  call}
}
\value{
entrez_summary output
}
\description{
Temp wrapper around entrez_summary because of errors in fetching data
}
