% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{extractAuthors}
\alias{extractAuthors}
\title{Get detailed MeSH info based on MeSH IDs or terms}
\usage{
extractAuthors(authorList, PMIDs, lastName, firstName)
}
\arguments{
\item{authorList}{vector of MeSH values to search for}

\item{PMIDs}{PMID vector corresponding to each entry in authorList XML object}

\item{lastName}{Last name}

\item{firstName}{First name. Add any middle initials the author uses to publish.
E.g. 'Joseph E Murray' --> 'Joseph E'. To be broader in case of multiple spellings,
only supply initials e.g. 'Joseph E Murray' --> 'JE'.}
}
\value{
a list with two data frames:
\itemize{
\item author: author name info
\item authors: all authors on articles
}
}
\description{
You should only use one parameter:
}
