% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{filter_affiliation}
\alias{filter_affiliation}
\title{Filter publicationDetails object based on author of interest affiliation match}
\usage{
filter_affiliation(publicationDetails, regex, includeMissing = F)
}
\arguments{
\item{publicationDetails}{Object generated by ncbi_publicationDetails}

\item{regex}{Search term (regular expression) to filter affiliations}

\item{includeMissing}{(Default = FALSE) Include results with missing
author affiliations}
}
\value{
Filtered publicationDetails object
}
\description{
This function is useful in case there are a lot of suprious matches that
would be too tedious to manage one by one
}
