% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{filter_PMID}
\alias{filter_PMID}
\title{Filter publicationDetails object based on author of interest affiliation match}
\usage{
filter_PMID(publicationDetails, PMIDs, remove = F, showWarnings = T)
}
\arguments{
\item{publicationDetails}{Object generated by ncbi_publicationDetails}

\item{PMIDs}{Vector of PMIDs to filter}

\item{remove}{(Default = F). If T the provided PMIDs will be removed instead of}

\item{showWarnings}{(Default = T). Display warnings if PMIDs don't match
filtered out}
}
\value{
Filtered publicationDetails object
}
\description{
This function is useful in case there are a lot of suprious matches that
would be too tedious to manage one by one
}
