% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFunctions.R
\name{generateProdFolder}
\alias{generateProdFolder}
\title{Generate Shiny App Production Folder
This will extract all relevant files needed to publish the shiny app on a server}
\usage{
generateProdFolder(folder, localDBs, overwrite = T)
}
\arguments{
\item{folder}{Folder where to create the ColabNet publication folder
(this folder will be generated if needed)}

\item{localDBs}{(Optional) Vector of any local databases to include in the app}

\item{overwrite}{Default = T. If set to F, existing files will generate error}
}
\value{
A ColabNet folder with all Shiny app files needed for production
}
\description{
Generate Shiny App Production Folder
This will extract all relevant files needed to publish the shiny app on a server
}
