% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePubmed.R
\name{getAuthor}
\alias{getAuthor}
\title{Get the best matching Pubmed author name}
\usage{
getAuthor(first, last)
}
\arguments{
\item{first}{First name}

\item{lastLast}{name}
}
\value{
Author name as string "lastName Initials" as used in Pubmed
}
\description{
Get the best matching Pubmed author name
}
