% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{lightenColour}
\alias{lightenColour}
\title{Make a hexadecimal colour lighter (go towards white)}
\usage{
lightenColour(hex, factor)
}
\arguments{
\item{hex}{A hexadecimal colour value}

\item{factor}{How much whiter to make the colour. 0 = no change, 1 = white}
}
\value{
A vector of "black" or "white" for each colour
}
\description{
Make a hexadecimal colour lighter (go towards white)
}
