% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{missingTreeNums}
\alias{missingTreeNums}
\title{Extract all missing Mesh tree numbers for leaf nodes to reach their root}
\usage{
missingTreeNums(treenums, checkNums = T)
}
\arguments{
\item{checkNums}{TRUE (Default) - If TRUE, will check if tree numbers are valid}

\item{known}{A vector of MeshTreeNumbers (e.g. N.06.850.290.200)}
}
\value{
vector of treenums representing missing nodes between leaves and root
}
\description{
This will allow to reconstruct the MeSH tree later by filling in gaps from
leaves to the root
}
