% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_admin.R
\name{mod_admin_server}
\alias{mod_admin_server}
\title{Admin module for Colabnet  - Server}
\usage{
mod_admin_server(id, conn)
}
\arguments{
\item{id}{ID for the module (match with UI)}

\item{conn}{A reactive conn object connected to the database}
}
\value{
Nothing
}
\description{
Admin module for Colabnet  - Server
}
