% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_admin.R
\name{mod_admin_ui}
\alias{mod_admin_ui}
\title{Admin module for Colabnet  - UI}
\usage{
mod_admin_ui(id)
}
\arguments{
\item{id}{ID for the module (match with server)}
}
\value{
UI object for database admin
}
\description{
Admin module for Colabnet  - UI
}
