% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_dbConnect_server}
\alias{mod_dbConnect_server}
\title{Module server to connect to an SQLite database in various ways}
\usage{
mod_dbConnect_server(id, localFolder, tempFolder)
}
\arguments{
\item{id}{ID name for the server}

\item{localFolder}{Folder with existing, permanent databases to provide}

\item{tempFolder}{Folder where temp (new and uploaded) databases live}
}
\value{
Reactive list variable with 4 items
\itemize{
\item dbPath: path to the database
\item dbName: (file) name of the database
\item dbCode: temporary code name of the database
\item dbType: method chosen from the menu
}
}
\description{
Module server to connect to an SQLite database in various ways
}
