% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_dbConnect_ui}
\alias{mod_dbConnect_ui}
\title{Module UI to connect to an SQLite database in various ways}
\usage{
mod_dbConnect_ui(id)
}
\arguments{
\item{id}{ID name for the module}
}
\value{
When inserted it will pop-up a modal window. There is no permanent UI
}
\description{
Module UI to connect to an SQLite database in various ways
}
