% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_dbSetup_ui}
\alias{mod_dbSetup_ui}
\title{Module UI to setup an SQLite database in various ways}
\usage{
mod_dbSetup_ui(id)
}
\arguments{
\item{id}{ID name for the module}
}
\value{
When inserted it will pop-up a modal window. There is no permanent UI
}
\description{
Note that the server function will automatically generate the modal pop-up
when initialised based on the settings / URL.
}
