% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meshFilter.R
\name{mod_meshFilter_server}
\alias{mod_meshFilter_server}
\title{MeSH filtering module - Server
This module allows the user to search and select specific MeSH terms
from the current database}
\usage{
mod_meshFilter_server(id, conn)
}
\arguments{
\item{id}{Module id, needs to be the same as the UI function}

\item{conn}{A reactive conn object as DB connection}
}
\value{
Server of the module that returns a reactive dataframe of MeSH terms
the user has selected
}
\description{
MeSH filtering module - Server
This module allows the user to search and select specific MeSH terms
from the current database
}
