% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meshFilter.R
\name{mod_meshFilter_ui}
\alias{mod_meshFilter_ui}
\title{MeSH filtering module - UI
This module allows the user to search and select specific MeSH terms
from the current database}
\usage{
mod_meshFilter_ui(id)
}
\arguments{
\item{id}{Module id, needs to be the same as the server function}
}
\value{
UI for the module
}
\description{
MeSH filtering module - UI
This module allows the user to search and select specific MeSH terms
from the current database
}
