% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_meshTree.R
\name{mod_meshTree_server}
\alias{mod_meshTree_server}
\title{MeSH Tree Shiny Module - Server}
\usage{
mod_meshTree_server(id, papermeshtree, conn)
}
\arguments{
\item{id}{ID for the module}

\item{papermeshtree}{A papermeshtree data frame}

\item{conn}{A reactive conn object as DB connection}
}
\value{
Server function for MeSH Tree. The output will return a reactive list:
\itemize{
\item mtrIDs: the mtrIDs currently in the treemap (based on filtering),
\item selected: the currently selected branch,
\item auIDs: the current author IDs
}
}
\description{
MeSH Tree Shiny Module - Server
}
