% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_meshTree_ui}
\alias{mod_meshTree_ui}
\title{MeSH Tree Shiny Module - UI}
\usage{
mod_meshTree_ui(id, limits = c(12, 10, 8, 6, 4))
}
\arguments{
\item{id}{ID for the module}

\item{limits}{Default = c(12, 10, 8, 6, 4). Values in the limits dropdown
menu. A default value for all will always be added}
}
\value{
UI for MeSH Tree
}
\description{
MeSH Tree Shiny Module - UI
}
