% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_author}
\alias{ncbi_author}
\title{Get the best matching Pubmed author name}
\usage{
ncbi_author(lastName, firstName, samples = 10, showWarnings = T)
}
\arguments{
\item{lastName}{Last name}

\item{firstName}{First name. Add any middle initials the author uses to publish.
E.g. 'Joseph E Murray' --> 'Joseph E'. To be broader in case of multiple spellings,
only supply initials e.g. 'Joseph E Murray' --> 'JE'.}

\item{samples}{(Default = 10). Number of papers to check for variations in
name spelling}

\item{showWarnings}{(Default = T) Show warnings}
}
\value{
A data frame with lastName, firstName, initials based on top hit
articles. If there are multiple name versions, they are grouped based on
similarity (using simpleText function)
}
\description{
Get the best matching Pubmed author name
}
