% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_meshInfo}
\alias{ncbi_meshInfo}
\title{Get detailed MeSH info based on MeSH IDs or terms}
\usage{
ncbi_meshInfo(terms = NULL, uid = NULL)
}
\arguments{
\item{terms}{vector of MeSH descriptor IDs, tree numbers, or other terms}

\item{uid}{MeSH uid (integer values)}
}
\value{
a list with two elements: meshTerms and meshTree
}
\description{
You should only use one parameter:
}
