% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_publicationDetails}
\alias{ncbi_publicationDetails}
\title{Get author names, papers, co-authors and affiliations for a specific researcher}
\usage{
ncbi_publicationDetails(PMIDs, n = -1)
}
\arguments{
\item{PMIDs}{Vector of PMIDs. Use ncbi_authorArticleList if you want to search by name}

\item{n}{(Default = -1 or all) Number of papers to fetch from Pubmed}
}
\value{
List with 5 data frames
author: Names the author used in papers
articles List of author's articles on Pubmed
coAuthors Co-authors on each paper
affiliations Affiliation of each author
meshDescriptors MeSH term descriptor associated with each paper
meshQualifiers MeSH term qualifier associated with each paper
}
\description{
Get author names, papers, co-authors and affiliations for a specific researcher
}
