% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooScore.R
\name{nodeEval}
\alias{nodeEval}
\title{Function to calculate node score and remainder to be pushed up to parent
This function is used inside of zooScore_tree()}
\usage{
nodeEval(lvlData, allID, carry, carryOrder, m)
}
\arguments{
\item{lvlData}{Data frame with tree info for the current level (see zooScore_tree)}

\item{allID}{A vector of all authors IDs (even if no papers for this node)}

\item{carry}{The papers carried over from child nodes. These were not used to
score at the previous level and can now be used at the current level with lower score}

\item{carryOrder}{Order of node IDs in carry, needed for proper matching}

\item{m}{Matrix (n x 2) with all pairs of auIDs}
}
\value{
A matrix (n x 5)
\itemize{
\item Cols 1 & 2 are all auID combinations
\item Cols 3 & 4 are remaining papers for au1 and au2, resp to be carrier over
\item Col 5 is the score for the author pair for this level (i.e. number of overlapping papers)
}
}
\description{
Function to calculate node score and remainder to be pushed up to parent
This function is used inside of zooScore_tree()
}
