% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{nodeSum}
\alias{nodeSum}
\title{Recursive algorithm to get sum of all values in a branch of a tree (by node)}
\usage{
nodeSum(node, parent, value)
}
\arguments{
\item{node}{ID of current node}

\item{parent}{ID of node's parent}

\item{value}{Value of the node}
}
\value{
Dataframe containing all info needed to build a plotly treemap
}
\description{
Recursive algorithm to get sum of all values in a branch of a tree (by node)
}
