% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooScore.R
\name{paperMat}
\alias{paperMat}
\title{Generate a matrix for a node in the graph with author paper info
This function is used inside of nodeEval()}
\usage{
paperMat(auID, allID, n, m)
}
\arguments{
\item{auID}{A vector of author IDs for the node of interest}

\item{allID}{A vector of all authors IDs (even if no papers for this node)}

\item{n}{The number of papers per author (in oder of auID)}

\item{m}{Matrix (n x 2) with all pairs of auIDs}
}
\value{
Paper count matrix for each author pair for the given node
}
\description{
Generate a matrix for a node in the graph with author paper info
This function is used inside of nodeEval()
}
