% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{paperMeshTree}
\alias{paperMeshTree}
\title{Expand a MeSH tree with author level info from papermesh}
\usage{
paperMeshTree(papermesh, meshtree)
}
\arguments{
\item{papermesh}{Data frame generated by dbPaperMesh()}

\item{meshtree}{Data frame generated by dbMeshTree()}
}
\value{
Dataframe containing the MeSH tree expanded with author details
}
\description{
Expand a MeSH tree with author level info from papermesh
}
