% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{papermeshtreeFromAuIDs}
\alias{papermeshtreeFromAuIDs}
\title{Generate a TreeMap from a list of author IDs}
\usage{
papermeshtreeFromAuIDs(auIDs, roots, conn)
}
\arguments{
\item{auIDs}{List of authors IDs}

\item{roots}{(Optional) Vector of single letter representing the tree roots
to include. If not specified, all categories are returned}

\item{conn}{Connection to a Colabnet database}
}
\value{
List with two elements
\itemize{
\item papermeshtree: Dataframe used to build a TreeMap
\item authors: Data frame with default author name and ID
}
}
\description{
Generate a TreeMap from a list of author IDs
}
