% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{plotDataFilter}
\alias{plotDataFilter}
\title{Filter treemap plot data based on limits}
\usage{
plotDataFilter(plotData, n)
}
\arguments{
\item{plotData}{A MeSH Tree plot data dataframe}

\item{n}{Number of top-scoring categories to show}
}
\value{
Data frame of plotData filtered by top-n categories per level
}
\description{
Filter treemap plot data based on limits
}
