% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotDiffTree}
\alias{plotDiffTree}
\title{Compare two author MeSH trees and find overlapping areas}
\usage{
plotDiffTree(difftree, colours, dbInfo)
}
\arguments{
\item{difftree}{a diffTree dataframe returned by diffTree()}

\item{colours}{(optional) A list of 3 colours
\itemize{
\item In case of comparing two authors, the first two colours represent the individual author's
MeSH terms and the last colour shared terms
\item In case of 2+, the first colour represents a term only belonging to a single author,
the next two colours form a scale between which the number of authors per MeSH term will
be interpolated. Min number of authors is colour 2, max is colour 3
}}

\item{dbInfo}{Path to the ColabNet database. Can be left blank if setup with dbConn()}
}
\value{
A plotly Treemap
}
\description{
Compare two author MeSH trees and find overlapping areas
}
