% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{profvisRender}
\alias{profvisRender}
\title{Wrapper function for profvis to work in Positron until issue resolved
https://github.com/posit-dev/positron/issues/3269}
\usage{
profvisRender(expr, folder = "local")
}
\arguments{
\item{expr}{R expression to evaluate with profvis}

\item{folder}{Default = "local". Folder where to store result files}
}
\value{
Run profvis and open the results in the browser
}
\description{
Wrapper function for profvis to work in Positron until issue resolved
https://github.com/posit-dev/positron/issues/3269
}
