% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrapePubmed.R
\name{simpleText}
\alias{simpleText}
\title{Simplify special characters in text for easier comparisons}
\usage{
simpleText(text)
}
\arguments{
\item{text}{}
}
\value{
simplified text
}
\description{
Simplify special characters in text for easier comparisons
}
