% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{tempFileCheck}
\alias{tempFileCheck}
\title{Function to check which temp files can be deleted based on time / size}
\usage{
tempFileCheck(folder, hours, totalSize, pattern = NULL)
}
\arguments{
\item{folder}{Temp folder}

\item{hours}{Number of hours after which files should be removed}

\item{totalSize}{Max size the temp folder can have in MB. This will mark files
for deletion, even if not expired (oldest first) until below limit}

\item{pattern}{(Optional) File name filter (e.g. extension)}
}
\value{
list with two elements
\itemize{
\item fileInfo: data frame with all files and annotation which ones to remove
\item summary: data frame with summary stats of the files
}
}
\description{
Function to check which temp files can be deleted based on time / size
}
