% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{textBW}
\alias{textBW}
\title{Check whether to use black or white text on a colour backgound}
\usage{
textBW(colours)
}
\arguments{
\item{colours}{A vector of colours}
}
\value{
A vector of "black" or "white" for each colour
}
\description{
Check whether to use black or white text on a colour backgound
}
