% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapBalance}
\alias{treemapBalance}
\title{Function to calculate values that will balance all rectangles of a Tree map
to be equal size}
\usage{
treemapBalance(id, parent)
}
\arguments{
\item{id}{Vector of node IDs in the tree}

\item{parent}{Vector of corresponding parent for each node in id (note that
the parent of the root ID should be NA)}
}
\value{
data frame with two columns
\itemize{
\item id: node ID
\item balanceVal: value that will balance the box size (only leaves get values)
}
}
\description{
Function to calculate values that will balance all rectangles of a Tree map
to be equal size
}
