% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapColour}
\alias{treemapColour}
\title{Generate colours for the treemap based on values}
\usage{
treemapColour(vals, minCol = "#f2f3f5", maxCol = "#4682B4", logScale = T)
}
\arguments{
\item{vals}{List of values to generate colours for}

\item{minCol}{Default = "#f2f3f5", Hex colour associated with lowest values
in the treemap}

\item{maxCol}{Default = "#4682B4", Hex colour associated with lowest values
in the treemap}

\item{logScale}{Default = TRUE. Log-scale the values for more gradual colour
transitions}
}
\value{
Vector of hexadecimal colour for each value
}
\description{
Generate colours for the treemap based on values
}
