% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treeMapComparison}
\alias{treeMapComparison}
\title{Generate colours for the treemap based on values}
\usage{
treeMapComparison(au1, au2, dbInfo)
}
\arguments{
\item{au1}{Author ID for first author}

\item{au2}{Author ID for second author}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe to build TreeMap
}
\description{
Generate colours for the treemap based on values
}
