% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapData}
\alias{treemapData}
\title{Build a full MeSH tree from papermesh data frame}
\usage{
treemapData(papermeshtree, colorComparison = T)
}
\arguments{
\item{papermeshtree}{Data frame generated by dbPaperMeshTree()}

\item{colorComparison}{(Default = TRUE). Colour comparisons of 2 authors}
}
\value{
Dataframe containing all info needed to build a plotly treemap
}
\description{
Build a full MeSH tree from papermesh data frame
}
