% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapFromAuIDs}
\alias{treemapFromAuIDs}
\title{Generate a TreeMap from a list of author IDs}
\usage{
treemapFromAuIDs(auIDs, roots, dbInfo)
}
\arguments{
\item{auIDs}{List of authors IDs}

\item{roots}{(Optional) Vector of single letter representing the tree roots
to include. If not specified, all categories are returned}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
List with two elements
\itemize{
\item treemapdata: Dataframe used to build a TreeMap
\item authors: Data frame with default author name and ID
}
}
\description{
Generate a TreeMap from a list of author IDs
}
