% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooScore.R
\name{zooScore}
\alias{zooScore}
\title{Calculate the zoo scores for a paper MeSH tree}
\usage{
zooScore(papermeshtree, roots = NULL)
}
\arguments{
\item{papermeshtree}{A Mesh Tree generated by paperMeshTree()}

\item{roots}{Optional. If set, a vector of roots in paperMeshTree to run the
algorithm for, otherwise all roots (i.e. trees) are used}
}
\value{
A data frame with 3 columns
\itemize{
\item Cols 1 & 2 are all auID combinations
\item Col 3 is the zoo score
}
}
\description{
Calculate the zoo scores for a paper MeSH tree
}
