% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zooScore.R
\name{zooScore_tree}
\alias{zooScore_tree}
\title{Calculate the zoo scores for all author combos in a single tree
This function is used inside of zooScore()}
\usage{
zooScore_tree(tree)
}
\arguments{
\item{tree}{A single Mesh Tree (one root) generated as part of the
paperMeshTree() function (generates all trees)}
}
\value{
A data frame with 3 columns
\itemize{
\item Cols 1 & 2 are all auID combinations
\item Col 3 is the zoo score
}
}
\description{
Calculate the zoo scores for all author combos in a single tree
This function is used inside of zooScore()
}
