% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{colabNet}
\alias{colabNet}
\title{Run the colabNet Shiny App}
\usage{
colabNet(colabNetDB, localFolder, tempFolder)
}
\arguments{
\item{colabNetDB}{Path to the ColabNet database (will be created if needed)}

\item{localFolder}{(Optional) Path to a folder with database to explore}

\item{tempFolder}{(Optional) Path to a folder where temp databases are stored

If local and temp folders are not set, they will be created in the default temp}
}
\value{
Start the Shiny app
}
\description{
Run the colabNet Shiny App
}
