% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{colabNet}
\alias{colabNet}
\title{Run the colabNet Shiny App}
\usage{
colabNet(colabNetDB)
}
\arguments{
\item{colabNetDB}{Path to the ColabNet database (will be created if needed)}
}
\value{
Start the Shiny app
}
\description{
Run the colabNet Shiny App
}
