% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddAuthorPublications}
\alias{dbAddAuthorPublications}
\title{Add authors to the database}
\usage{
dbAddAuthorPublications(authorPublications, dbPath)
}
\arguments{
\item{authorPublications}{List of data frames geneated by ncbi_authorPublications()}

\item{dbPath}{(optional if dbSetup() has been run) Path to the ColabNet database}
}
\value{
Data is inserted into the DB returning a list the author's aritcles (arID and PMID)
}
\description{
Add authors to the database
}
