% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddAuthorPublications}
\alias{dbAddAuthorPublications}
\title{Add authors to the database}
\usage{
dbAddAuthorPublications(
  authorPublications,
  matchOnFirst = F,
  flagUpdate = T,
  dbInfo
)
}
\arguments{
\item{authorPublications}{List of data frames geneated by ncbi_publicationDetails()}

\item{flagUpdate}{Default= T, set an update flag in the DB so the app will refresh once completed}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Data is inserted into the DB returning a list the author's aritcles (arID and PMID)
}
\description{
Add authors to the database
}
