% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddAuthors}
\alias{dbAddAuthors}
\title{Add authors to the database}
\usage{
dbAddAuthors(authors, dbInfo, commit = T)
}
\arguments{
\item{authors}{The coAuthors data frame generated by ncbi_authorPublications}

\item{dbInfo}{Connection to a Colabnet database}

\item{commit}{(Default = TRUE). Commit changes to the database automatically}
}
\value{
Data frame with new, updated and existing author info
}
\description{
Add authors to the database
}
