% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddAuthors}
\alias{dbAddAuthors}
\title{Add authors to the database}
\usage{
dbAddAuthors(authors, dbInfo)
}
\arguments{
\item{authors}{The coAuthors data frame generated by ncbi_authorPublications}

\item{dbInfo}{(optional if dbSetup() has been run) Path to the ColabNet database
or existing connection}
}
\value{
Data frame with new, updated and existing author info
}
\description{
Add authors to the database
}
