% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddAuthors}
\alias{dbAddAuthors}
\title{Add authors to the database}
\usage{
dbAddAuthors(authors, dbPath)
}
\arguments{
\item{authors}{A data frame with columns lastName, firstName, initials and collectiveName generated by
one of the MeSH functions}

\item{dbPath}{(optional if dbSetup() has been run) Path to the ColabNet database}
}
\value{
Data frame with new, updated and existing author info
}
\description{
Add authors to the database
}
