% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbAddMesh}
\alias{dbAddMesh}
\title{Insert MeSH info into the ColabNet database}
\usage{
dbAddMesh(values, type, dbPath)
}
\arguments{
\item{values}{vector of MeSH values to search for}

\item{type}{The type needs to be 'meshui' (MeSH ui), 'treenum' (tree number) or 'uid' (MeSH Entrez uid)}

\item{dbPath}{(optional if dbSetup() has been run) Path to the ColabNet database}
}
\value{
data frame with all uid and meshui inserted into the database (including intermediate tree nodes)
}
\description{
Insert MeSH info into the ColabNet database
}
