% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbDeleteArticle}
\alias{dbDeleteArticle}
\title{Delete specified articles and any associated authors / author names
Useful when an article has been incorrectly associated with an author for
example because of name overlap}
\usage{
dbDeleteArticle(arIDs, dbInfo)
}
\arguments{
\item{arIDs}{Vector of arIDs to delete (as found in article table)}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe of articles and authors/author names that have been deleted
}
\description{
Delete specified articles and any associated authors / author names
Useful when an article has been incorrectly associated with an author for
example because of name overlap
}
