% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbDeleteAuthors}
\alias{dbDeleteAuthors}
\title{Remove authors and their articles from the database}
\usage{
dbDeleteAuthors(auIDs, dbInfo, commit = T)
}
\arguments{
\item{auIDs}{auIDs to remove. Co-authors who are not and authorOfInterest and
do not appear in other non-removed articles, are removed too}

\item{dbInfo}{Connection to a Colabnet database}

\item{commmit}{(Default = TRUE). Commit changes to the database automatically}
}
\value{
data frame with all auID and arID that were removed
}
\description{
Remove authors and their articles from the database
}
