% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbFlagUpdate}
\alias{dbFlagUpdate}
\title{Tell the colabNet app an update has been made and it needs to refresh}
\usage{
dbFlagUpdate(action, dbInfo)
}
\arguments{
\item{action}{(Integer) The code of the action
\itemize{
\item 0 = new database init
\item 1 = add data
\item 2 = delete data
}}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Bool to indicate success
}
\description{
Tell the colabNet app an update has been made and it needs to refresh
}
