% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Get a ColabNet database connection}
\usage{
dbGetConn(dbInfo)
}
\arguments{
\item{dbInfo}{(optional if dbSetup() has been run) Path to the ColabNet database
or existing connection. If the dbSetup() has been run the assigned database will be used}
}
\value{
Connection to the ColabNet database
}
\description{
Get a ColabNet database connection
}
