% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBoperations.R
\name{dbGetConn}
\alias{dbGetConn}
\title{Add authors to the database}
\usage{
dbGetConn(path, newDBMsg = T, checkSchema = F)
}
\arguments{
\item{path}{Path to ColabNet Database. If DB does not exist it will be created}

\item{newDBMsg}{(Default, TRUE) Output a message when a new database was created}

\item{checkSchema}{(Default, FALSE) Check the schema of an existing datbase against the reference}
}
\value{
A connection to the ColabNet datbase
}
\description{
Add authors to the database
}
