% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{dbMeshTree}
\alias{dbMeshTree}
\title{Build a full MeSH tree from a papermesh data frame}
\usage{
dbMeshTree(papermesh, dbInfo)
}
\arguments{
\item{papermesh}{Data frame generated by dbPaperMesh()}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe containing the MeSH tree based on papermesh info
This will also contain parentIDs, branchIDs and parentBranchIDs
the tree root name and treemapVals
}
\description{
Build a full MeSH tree from a papermesh data frame
}
