% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{dbTreeFromMesh}
\alias{dbTreeFromMesh}
\title{Build a MeSH Tree from a list of MeSH IDs
Missing nodes will be filled in and the tree will be built until the root}
\usage{
dbTreeFromMesh(uids, dbInfo)
}
\arguments{
\item{uids}{Vector of Mesh uid to buidll the tree}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe with tree info (mtrID, uid, meshui, treenum)
}
\description{
Build a MeSH Tree from a list of MeSH IDs
Missing nodes will be filled in and the tree will be built until the root
}
