% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{missingTreeNums}
\alias{missingTreeNums}
\title{Extract all missing Mesh tree numbers for leaf nodes to reach their root}
\usage{
missingTreeNums(treenums, includeOriginal = F, includeRoots = F)
}
\arguments{
\item{treenums}{A vector of MeshTreeNumbers (e.g. N.06.850.290.200).
Use checkTreeNums() is you want to make sure the numbers are in valid format}

\item{includeOriginal}{Default = FALSE. If TRUE, the provided treenums will be
part of the result}

\item{includeRoots}{Default = FALSE. If TRUE, the provided root (category)
letter will be part of the result}
}
\value{
vector of treenums representing missing nodes between leaves and root
}
\description{
This will allow to reconstruct the MeSH tree later by filling in gaps from
leaves to the root
}
