% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_dbSetup_ui}
\alias{mod_dbSetup_ui}
\title{Module UI to setup an SQLite database in various ways}
\usage{
mod_dbSetup_ui(id, download = "button")
}
\arguments{
\item{id}{ID name for the module}

\item{download}{Default = "button". Element to download the current database.
Choose from button, link or none}
}
\value{
Module UI (with download element if not none)
}
\description{
Note that the server function will automatically generate the modal pop-up
when initialised based on the settings / URL.
}
