% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{mod_meshTree_server}
\alias{mod_meshTree_server}
\title{MeSH Tree Shiny Module - Server}
\usage{
mod_meshTree_server(id, plotData)
}
\arguments{
\item{id}{ID for the module}

\item{plotData}{A MeshTree plot data data frame}
}
\value{
Server function for MeSH Tree. The output will retrun the currently
selected branchID
}
\description{
MeSH Tree Shiny Module - Server
}
