% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_author}
\alias{ncbi_author}
\title{Get the best matching Pubmed author name}
\usage{
ncbi_author(firstName, lastName)
}
\arguments{
\item{firstName}{First name}

\item{lastName}{Last name}
}
\value{
Author name as string "lastName Initials" as used in Pubmed
}
\description{
Get the best matching Pubmed author name
}
