% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_author}
\alias{ncbi_author}
\title{Get the best matching Pubmed author name}
\usage{
ncbi_author(lastName, firstName)
}
\arguments{
\item{lastName}{Last name}

\item{firstName}{First name}
}
\value{
A data frame with lastName, firstName, initials based on top hit article
}
\description{
Get the best matching Pubmed author name
}
