% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_authorArticleList}
\alias{ncbi_authorArticleList}
\title{Get basic article info for an author if name is likely unique}
\usage{
ncbi_authorArticleList(
  lastName,
  firstName,
  PMIDs,
  returnHistory = F,
  PMIDonly = F,
  simpletext = T,
  stopFetching = 1000
)
}
\arguments{
\item{lastName}{Author last name}

\item{firstName}{First name. Add any middle initials the author uses to publish.
E.g. 'Joseph E Murray' --> 'Joseph E'. To be broader in case of multiple spellings,
only supply initials e.g. 'Joseph E Murray' --> 'JE'.}

\item{PMIDs}{(optional) If set limit the search to the PMIDs for the given author}

\item{returnHistory}{(Default = False). If true, will return rentrez web history object}

\item{PMIDonly}{Return only valid PMID, not the data table}

\item{stopFetching}{(Default = 1000) If more than this number of articles are
found, it is very likely that results for multiple authors with the same name are
found. In this case is might be better to manually provide the PMID instead}
}
\value{
List with 4 elements
\itemize{
\item success: TRUE if success; If FALSE, number of matches > stopFetching, nothing returned
\item n: Total number of articles found (returned even when stopFetching reached)
\item PMIDs: when not stopFetching
\item articles: Data frame with basic article info (when PMIDonly = F and not stopFetching)
\item history: If returnHistory = T: History object that can be used for subsequent entrez
calls (for large datasets), otherwise NA
}
}
\description{
Get basic article info for an author if name is likely unique
}
