% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_authorArticleList}
\alias{ncbi_authorArticleList}
\title{Get basic article info for an author if name is likely unique}
\usage{
ncbi_authorArticleList(
  lastName,
  firstName,
  initials,
  searchInitials = T,
  PMIDonly = F,
  stopFetching = 100
)
}
\arguments{
\item{lastName}{Author last name}

\item{firstName}{Author first name}

\item{initials}{Author initials (ignored when searchInitials = F)}

\item{searchInitials}{Default = T. Include additional search on initials.
This is helpful if author has used multiple versions of first name}

\item{PMIDonly}{Last name + initials, as generated by ncbi_author}

\item{stopFetching}{(Default = 100) If more than this number of articles are
found, it might be that results for multiple authors with the same name are
found. In this case is might be better to manually provide the PMID instead}
}
\value{
List with 4 elements
statusCode:
\itemize{
\item 0: too many hits with first name and initials (defined by stopFetching).
only total count returned
\item 1: too many hits with initials or searchInitials = F.
First name hits returned only
\item 2: hits with first name and initials
\item n: Total number of articles found (returned even when stopFetching reached)
\item PMIDs: when not stopFetching
\item articles: Data frame with basic article info (when PMIDonly = F and not stopFetching)
}
}
\description{
Get basic article info for an author if name is likely unique
}
