% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_authorArticleList}
\alias{ncbi_authorArticleList}
\title{Get basic article info for an author if name is likely unique}
\usage{
ncbi_authorArticleList(author, PMIDonly = F, stopFetching = 100)
}
\arguments{
\item{author}{Last name + initials, as generated by ncbi_author}

\item{stopFetching}{(Default = 100) If more than this number of articles are
found, it might be that results for multiple authors with the same name are
found. In this case is might be better to manually provide the PMID instead}
}
\value{
List with three elements
\itemize{
\item Whether the search was successful or stopFetching was reached
\item Total number of articles found (returned even when stopFetching reached)
\item Data frame with basic article info (on when not stopFetching)
}
}
\description{
Get basic article info for an author if name is likely unique
}
