% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCalls.R
\name{ncbi_meshInfo}
\alias{ncbi_meshInfo}
\title{Get detailed MeSH info based on MeSH IDs or terms}
\usage{
ncbi_meshInfo(values, type = c("meshui", "treenum", "uid"))
}
\arguments{
\item{values}{vector of MeSH values to search for}

\item{type}{The type needs to be meshui (MeSH ui) , treenum (tree number) or uid (MeSH Entrez uid)}
}
\value{
a list with two elements: meshTerms and meshTree
}
\description{
You should only use one parameter:
}
