% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{plotDiffTree}
\alias{plotDiffTree}
\title{Compare two author MeSH trees and find overlapping areas}
\usage{
plotDiffTree(diffTree, colours = c("#69BE28", "#3DB7E4", "#FF8849"))
}
\arguments{
\item{diffTree}{a diffTree dataframe returned by diffTree()}

\item{colours}{A list of 3 colours for colouring shared,
author 1 unique and author 2 unique MeSH terms}
}
\value{
A plotly Treemap
}
\description{
Compare two author MeSH trees and find overlapping areas
}
