% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treeMapComparison}
\alias{treeMapComparison}
\title{Generate a treemap that compares two authors}
\usage{
treeMapComparison(au1, au2, roots, dbInfo)
}
\arguments{
\item{au1}{Author ID for first author}

\item{au2}{Author ID for second author}

\item{roots}{(Optional) Vector of single letter representing the tree roots
to include. If not specified, all categories are returned}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe to build TreeMap
}
\description{
Generate a treemap that compares two authors
}
