% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapData}
\alias{treemapData}
\title{Build a full MeSH tree from papermesh data frame}
\usage{
treemapData(papermeshtree)
}
\arguments{
\item{papermeshtree}{Data frame generated by dbPaperMeshTree()}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
Dataframe containing all info needed to build a plotly treemap
}
\description{
Build a full MeSH tree from papermesh data frame
}
