% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{treemapFromAuIDs}
\alias{treemapFromAuIDs}
\title{Generate a TreeMap from a list of author IDs}
\usage{
treemapFromAuIDs(auIDs, dbInfo)
}
\arguments{
\item{auIDs}{List of authors IDs}

\item{dbInfo}{(optional if dbSetup() has been run)
Path to the ColabNet database or existing connection}
}
\value{
List with two elements
\itemize{
\item treemapdata: Dataframe used to build a TreeMap
\item authors: Data frame with default author name and ID
}
}
\description{
Generate a TreeMap from a list of author IDs
}
